Move AtomStorage, AtomTrieNode and AtomIndex into separate modules
